Feat: RecordMap.vue -> features/ui 로 기능분리
